Added pattern support to captures
This is used by the official C# TM language file, so it presumably works for some engines. Even if it doesn't work for all engines, we should still support it for those that do.

Here is the sample from the official DotNet group C# TM Language spec: https://github.com/dotnet/csharp-tmLanguage/blob/9d446ab1fd31271c7c935069669e8504ddddb8ad/grammars/csharp.tmLanguage#L7330-L7352

The PLIST version of the file got regenerated with this very plugin, and for some reason lost track of all its spacing. It was a single line added to the JSON version of the TM language.